refactor: trade 관련 interface 분리
